+++ livecomponents/types.py
@@ -2,7 +2,7 @@ from typing import TypeVar
 
 from pydantic import BaseModel, ConfigDict
 
-from livecomponents.utils import LiveComponentsPath
+from livecomponents.utils import LiveComponentsPath, get_ancestor_id
 
 State = TypeVar("State")
 
@@ -10,6 +10,21 @@ State = TypeVar("State")
 class StateAddress(BaseModel):
     session_id: str
     component_id: str
+
+    def find_ancestor(self, ancestor_type: str) -> "StateAddress | None":
+        """Find the closest ancestor of the given type."""
+        ancestor_component_id = get_ancestor_id(self.component_id, ancestor_type)
+        if not ancestor_component_id:
+            return None
+        return StateAddress(
+            session_id=self.session_id, component_id=ancestor_component_id
+        )
+
+    def must_find_ancestor(self, ancestor_type: str) -> "StateAddress":
+        ancestor = self.find_ancestor(ancestor_type)
+        if not ancestor:
+            raise ValueError(f"Ancestor {ancestor_type} not found")
+        return ancestor
 
     def get_parent(self) -> "StateAddress | None":
         """Returns the parent of this component or None.